# openapi.json
{
  "components": {
    "schemas": {
...
      "EvalOut": {
        "properties": {
          "cost_per_doc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Per Doc"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "doc_count": {
            "title": "Doc Count",
            "type": "integer"
          },
          "doc_exact_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Exact Rate"
          },
          "effort": {
            "title": "Effort",
            "type": "string"
          },
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "fewshot_k": {
            "title": "Fewshot K",
            "type": "integer"
          },
          "field_order": {
            "items": {
              "type": "string"
            },
            "title": "Field Order",
            "type": "array"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lenient_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lenient Accuracy"
          },
          "lists": {
            "additionalProperties": true,
            "title": "Lists",
            "type": "object"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "overall_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Accuracy"
          },
          "p50_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "P50 Ms"
          },
          "p95_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Ms"
          },
          "per_field": {
            "additionalProperties": true,
            "title": "Per Field",
            "type": "object"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "reps": {
            "title": "Reps",
            "type": "integer"
          },
          "schema_id": {
            "format": "uuid",
            "title": "Schema Id",
            "type": "string"
          },
          "split": {
            "title": "Split",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "succeeded": {
            "title": "Succeeded",
            "type": "integer"
          },
          "truncated": {
            "title": "Truncated",
            "type": "integer"
          },
          "use_ocr_text": {
            "title": "Use Ocr Text",
            "type": "boolean"
          },
          "value_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Accuracy"
          }
        },
        "required": [
          "id",
          "schema_id",
          "prompt_version",
          "model",
          "effort",
          "fewshot_k",
          "use_ocr_text",
          "provider",
          "split",
          "mode",
          "notes",
          "status",
          "doc_count",
          "reps",
          "succeeded",
          "truncated",
          "failed",
          "per_field",
          "lists",
          "errors",
          "created_at"
        ],
        "title": "EvalOut",
        "type": "object"
      },
      "EvalRequest": {
        "properties": {
          "effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effort"
          },
          "fewshot_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fewshot K"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "mode": {
            "default": "sync",
            "title": "Mode",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "prompt": {
            "default": "v1",
            "title": "Prompt",
            "type": "string"
          },
          "reps": {
            "default": 1,
            "title": "Reps",
            "type": "integer"
          },
          "schema_name": {
            "default": "receipt",
            "title": "Schema Name",
            "type": "string"
          },
          "split": {
            "default": "test",
            "title": "Split",
            "type": "string"
          },
          "use_ocr_text": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Ocr Text"
          }
        },
        "title": "EvalRequest",
        "type": "object"
      },
      "EvalResultOut": {
        "properties": {
          "all_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "All Correct"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "document_name": {
            "title": "Document Name",
            "type": "string"
          },
          "extraction_run_id": {
            "format": "uuid",
            "title": "Extraction Run Id",
            "type": "string"
          },
          "outcomes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcomes"
          },
          "rep": {
            "title": "Rep",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "document_name",
          "extraction_run_id",
          "rep",
          "status"
        ],
        "title": "EvalResultOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaskAccepted": {
        "properties": {
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id"
        ],
        "title": "TaskAccepted",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Schema-driven document extraction with a review UI and per-field evals.",
    "title": "fieldwise",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evals": {
      "get": {
        "operationId": "list_evals_evals_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvalOut"
                  },
                  "title": "Response List Evals Evals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evals",
        "tags": [
          "evals"
        ]
      },
      "post": {
        "operationId": "start_eval_evals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Eval",
        "tags": [
          "evals"
        ]
      }
    },
    "/evals/compare": {
      "get": {
        "operationId": "compare_evals_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": true,
            "schema": {
              "items": {
                "format": "uuid",
                "type": "string"
              },
              "title": "Ids",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compare Evals Compare Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "evals"
        ]
      }
    },
    "/evals/{eval_id}": {
      "get": {
        "operationId": "get_eval_evals__eval_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "eval_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Eval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Eval",
        "tags": [
          "evals"
        ]
      }
    },
    "/evals/{eval_id}/results": {
      "get": {
        "operationId": "get_results_evals__eval_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "eval_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Eval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvalResultOut"
                  },
                  "title": "Response Get Results Evals  Eval Id  Results Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Results",
        "tags": [
          "evals"
        ]
      }
    },
...
  }
}